pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["alex.lin"]
readme = "README.md"

[tool.poetry.dependencies]
django = "^5.1"
djangorestframework = "^3.15.2"
django-cors-headers = "4.3.1"
drf-spectacular = "^0.27.2"
gunicorn = "22.0.0"
psycopg2-binary = "2.9.9"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
python = "^3.12"
whitenoise = "6.6.0"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
